Brok up beams_methods.cpp into several smaller files
