+++ src/restruct_poc/AssembleResidualVector.hpp
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <Kokkos_Core.hpp>
+#include <Kokkos_Profiling_ScopedRegion.hpp>
+
+#include "beams.hpp"
+#include "types.hpp"
+#include "IntegrateResidualVector.hpp"
+
+namespace openturbine {
+
+inline void AssembleResidualVector(Beams& beams, View_N residual_vector) {
+    auto region = Kokkos::Profiling::ScopedRegion("Assemble Residual");
+    Kokkos::parallel_for(
+        "IntegrateResidualVector", beams.num_nodes,
+        IntegrateResidualVector{
+            beams.node_state_indices, beams.node_FE, beams.node_FI, beams.node_FG, beams.node_FX,
+            residual_vector}
+    );
+}
+
+}
