+++ src/restruct_poc/UpdateState.hpp
@@ -1,77 +1,37 @@
-#include "beams_methods.hpp"
+#pragma once
 
-#include <array>
+#include <Kokkos_Core.hpp>
 #include <Kokkos_Profiling_ScopedRegion.hpp>
 
-
-#include "CalculateCuu.hpp"
-#include "CalculateForceFC.hpp"
-#include "CalculateForceFD.hpp"
-#include "CalculateGravityForce.hpp"
-#include "CalculateGyroscopicMatrix.hpp"
-#include "CalculateInertiaStiffnessMatrix.hpp"
-#include "CalculateInertialForces.hpp"
-#include "CalculateJacobian.hpp"
-#include "CalculateMassMatrixComponents.hpp"
-#include "CalculateMuu.hpp"
-#include "CalculateNodeForces.hpp"
+#include "types.hpp"
+#include "beams.hpp"
+#include "UpdateNodeState.hpp"
+#include "InterpolateQPState.hpp"
+#include "InterpolateQPVelocity.hpp"
+#include "InterpolateQPAcceleration.hpp"
 #include "CalculateOuu.hpp"
 #include "CalculatePuu.hpp"
 #include "CalculateQuu.hpp"
 #include "CalculateRR0.hpp"
+#include "CalculateMuu.hpp"
+#include "CalculateCuu.hpp"
 #include "CalculateStrain.hpp"
+#include "CalculateMassMatrixComponents.hpp"
 #include "CalculateTemporaryVariables.hpp"
-#include "IntegrateElasticStiffnessMatrix.hpp"
-#include "IntegrateMatrix.hpp"
-#include "IntegrateResidualVector.hpp"
-#include "InterpolateQPAcceleration.hpp"
-#include "InterpolateQPPosition.hpp"
-#include "InterpolateQPRotation.hpp"
-#include "InterpolateQPRotationDerivative.hpp"
-#include "InterpolateQPState.hpp"
-#include "InterpolateQPVelocity.hpp"
+#include "CalculateForceFC.hpp"
+#include "CalculateForceFD.hpp"
+#include "CalculateInertialForces.hpp"
+#include "CalculateGravityForce.hpp"
+#include "CalculateGyroscopicMatrix.hpp"
+#include "CalculateInertiaStiffnessMatrix.hpp"
+#include "CalculateNodeForces.hpp"
+
+
 
 namespace openturbine {
 
-struct UpdateNodeState {
-    Kokkos::View<int*> node_state_indices;
-    View_Nx7 node_u;
-    View_Nx6 node_u_dot;
-    View_Nx6 node_u_ddot;
-
-    View_Nx7::const_type Q;
-    View_Nx6::const_type V;
-    View_Nx6::const_type A;
-
-    KOKKOS_FUNCTION
-    void operator()(int i) const {
-        auto j = node_state_indices(i);
-        for (int k = 0; k < kLieGroupComponents; k++) {
-            node_u(i, k) = Q(j, k);
-        }
-        for (int k = 0; k < kLieAlgebraComponents; k++) {
-            node_u_dot(i, k) = V(j, k);
-        }
-        for (int k = 0; k < kLieAlgebraComponents; k++) {
-            node_u_ddot(i, k) = A(j, k);
-        }
-    }
-
-    KOKKOS_FUNCTION
-    void operator()(const int i, const int k) const {
-        auto j = node_state_indices(i);
-        node_u(i, k) = Q(j, k);
-        if (k < kLieAlgebraComponents) {
-            node_u_dot(i, k) = V(j, k);
-            node_u_ddot(i, k) = A(j, k);
-        }
-    }
-};
-
-// Update node states (displacement, velocity, acceleration) and interpolate to quadrature points
-void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
+inline void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
     auto region = Kokkos::Profiling::ScopedRegion("Update State");
-    // Copy displacement, velocity, and acceleration to nodes
     Kokkos::parallel_for(
         "UpdateNodeState", Kokkos::MDRangePolicy{{0, 0}, {beams.num_nodes, kLieGroupComponents}},
         UpdateNodeState{
@@ -85,7 +45,6 @@ void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
         }
     );
 
-    // Interpolate node state to quadrature points
     Kokkos::parallel_for(
         "InterpolateQpU", Kokkos::MDRangePolicy{{0, 0}, {beams.num_elems, beams.max_elem_qps}},
         InterpolateQPU{beams.elem_indices, beams.shape_interp, beams.node_u, beams.qp_u}
@@ -137,7 +96,6 @@ void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
         }
     );
 
-    // Calculate RR0 matrix
     Kokkos::parallel_for(
         "CalculateRR0", beams.num_qps,
         CalculateRR0{
@@ -147,19 +105,16 @@ void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
         }
     );
 
-    // Calculate Muu matrix
     Kokkos::parallel_for(
         "CalculateMuu", beams.num_qps,
         CalculateMuu{beams.qp_RR0, beams.qp_Mstar, beams.qp_Muu, beams.M_6x6}
     );
 
-    // Calculate Cuu matrix
     Kokkos::parallel_for(
         "CalculateCuu", beams.num_qps,
         CalculateCuu{beams.qp_RR0, beams.qp_Cstar, beams.qp_Cuu, beams.M_6x6}
     );
 
-    // Calculate strain
     Kokkos::parallel_for(
         "CalculateStrain", beams.num_qps,
         CalculateStrain{
@@ -171,7 +126,6 @@ void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
         }
     );
 
-    // Calculate Forces
     Kokkos::parallel_for(
         "CalculateMassMatrixComponents", beams.num_qps,
         CalculateMassMatrixComponents{beams.qp_Muu, beams.qp_eta, beams.qp_rho, beams.qp_eta_tilde}
@@ -228,7 +182,6 @@ void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
             beams.M9_3x3, beams.qp_Kuu}
     );
 
-    // Calculate nodal force vectors
     Kokkos::parallel_for(
         "CalculateNodeForces",
         Kokkos::MDRangePolicy{{0, 0}, {beams.num_elems, beams.max_elem_nodes}},
@@ -239,94 +192,4 @@ void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
     );
 }
 
-void AssembleResidualVector(Beams& beams, View_N residual_vector) {
-    auto region = Kokkos::Profiling::ScopedRegion("Assemble Residual");
-    Kokkos::parallel_for(
-        "IntegrateResidualVector", beams.num_nodes,
-        IntegrateResidualVector{
-            beams.node_state_indices, beams.node_FE, beams.node_FI, beams.node_FG, beams.node_FX,
-            residual_vector}
-    );
 }
-
-void AssembleMassMatrix(Beams& beams, View_NxN M) {
-    auto region = Kokkos::Profiling::ScopedRegion("Assemble Mass Matrix");
-    Kokkos::parallel_for(
-        "IntegrateMatrix",
-        Kokkos::MDRangePolicy{
-            {0, 0, 0, 0},
-            {beams.num_elems, beams.max_elem_nodes, beams.max_elem_nodes, beams.max_elem_qps}},
-        IntegrateMatrix{
-            beams.elem_indices,
-            beams.node_state_indices,
-            beams.qp_weight,
-            beams.qp_jacobian,
-            beams.shape_interp,
-            beams.qp_Muu,
-            M,
-        }
-    );
-}
-
-void AssembleGyroscopicInertiaMatrix(Beams& beams, View_NxN G) {
-    auto region = Kokkos::Profiling::ScopedRegion("Assemble Gyroscopic Inertia Matrix");
-    Kokkos::parallel_for(
-        "IntegrateMatrix",
-        Kokkos::MDRangePolicy{
-            {0, 0, 0, 0},
-            {beams.num_elems, beams.max_elem_nodes, beams.max_elem_nodes, beams.max_elem_qps}},
-        IntegrateMatrix{
-            beams.elem_indices,
-            beams.node_state_indices,
-            beams.qp_weight,
-            beams.qp_jacobian,
-            beams.shape_interp,
-            beams.qp_Guu,
-            G,
-        }
-    );
-}
-
-void AssembleInertialStiffnessMatrix(Beams& beams, View_NxN K) {
-    auto region = Kokkos::Profiling::ScopedRegion("Assemble Inertial Stiffness Matrix");
-    Kokkos::parallel_for(
-        "IntegrateMatrix",
-        Kokkos::MDRangePolicy{
-            {0, 0, 0, 0},
-            {beams.num_elems, beams.max_elem_nodes, beams.max_elem_nodes, beams.max_elem_qps}},
-        IntegrateMatrix{
-            beams.elem_indices,
-            beams.node_state_indices,
-            beams.qp_weight,
-            beams.qp_jacobian,
-            beams.shape_interp,
-            beams.qp_Kuu,
-            K,
-        }
-    );
-}
-
-void AssembleElasticStiffnessMatrix(Beams& beams, View_NxN K) {
-    auto region = Kokkos::Profiling::ScopedRegion("Assemble Elastic Stiffness Matrix");
-    Kokkos::parallel_for(
-        "IntegrateElasticStiffnessMatrix",
-        Kokkos::MDRangePolicy{
-            {0, 0, 0, 0},
-            {beams.num_elems, beams.max_elem_nodes, beams.max_elem_nodes, beams.max_elem_qps}},
-        IntegrateElasticStiffnessMatrix{
-            beams.elem_indices,
-            beams.node_state_indices,
-            beams.qp_weight,
-            beams.qp_jacobian,
-            beams.shape_interp,
-            beams.shape_deriv,
-            beams.qp_Puu,
-            beams.qp_Cuu,
-            beams.qp_Ouu,
-            beams.qp_Quu,
-            K,
-        }
-    );
-}
-
-}  // namespace openturbine
