+++ src/restruct_poc/beams.hpp
@@ -2,4 +2,3 @@
 
 #include "beams_data.hpp"
 #include "beams_input.hpp"
-#include "beams_methods.hpp"
