+++ tests/unit_tests/restruct_poc/test_beams.cpp
@@ -4,6 +4,12 @@
 
 #include "src/restruct_poc/beams.hpp"
 #include "src/restruct_poc/state.hpp"
+#include "src/restruct_poc/UpdateState.hpp"
+#include "src/restruct_poc/AssembleResidualVector.hpp"
+#include "src/restruct_poc/AssembleMassMatrix.hpp"
+#include "src/restruct_poc/AssembleGyroscopicInertiaMatrix.hpp"
+#include "src/restruct_poc/AssembleInertialStiffnessMatrix.hpp"
+#include "src/restruct_poc/AssembleElasticStiffnessMatrix.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
 
 namespace openturbine::restruct_poc::tests {
